==> 13_microdiversity.Rmd <==
# Microdiversity

```{r microdiversity_core,comment="", message=FALSE, warning=FALSE}
load("data/gut/data.Rdata")

samples <- c("EHI02067","EHI02065","EHI02617","EHI01968","EHI02630","EHI02639","EHI02607","EHI02582","EHI02612","EHI02619","EHI01985","EHI01966","EHI02615","EHI02695","EHI02592","EHI02585","EHI02602","EHI02632","EHI02698","EHI02697","EHI02624","EHI02693","EHI02088","EHI01995","EHI02097","EHI01970","EHI02633","EHI02598","EHI01989","EHI02079","EHI02696","EHI01967","EHI01992","EHI02105","EHI01979","EHI01982","EHI02625","EHI02603","EHI02584","EHI02587","EHI02694")

ani_files <- list.files(
  path      = "data/microdiversity/core",
  pattern   = "_population_ani\\.tsv$",
  full.names= TRUE
) %>%
  set_names(basename(.) %>% str_remove("_population_ani\\.tsv$"))

ani_list <- lapply(ani_files, function(file) {
  read_tsv(file,comment = "#") %>% 
  mutate(SampleID=samples) %>% 
  column_to_rownames(var="SampleID") %>% 
  set_names(samples) %>%
  select(where(~ !all(is.na(.)))) %>%
  drop_na()
}) %>%
  setNames(names(ani_files))

ani_pairwise <- ani_list %>%
  imap_dfr(~ {
    mat    <- .x
    genome <- .y
    as.data.frame(mat) %>%
      rownames_to_column("sample.x") %>%
      pivot_longer(-sample.x,
                   names_to  = "sample.y",
                   values_to = "distance") %>%
      mutate(genome = genome) %>%
      select(genome, sample.x, sample.y, distance)
  })

ani_environments <- ani_pairwise %>% 
  filter(sample.x != sample.y) %>% 
  inner_join(sample_metadata %>% select(sample,environment), by=join_by("sample.x"=="sample")) %>% 
  inner_join(sample_metadata %>% select(sample,environment), by=join_by("sample.y"=="sample")) %>% 
  mutate(environment=case_when(
            environment.x == "low" & environment.y == "high" ~ "different",
            environment.x == "high" & environment.y == "low" ~ "different",
            environment.x == "low" & environment.y == "low" ~ "low",
            environment.x == "high" & environment.y == "high" ~ "high",
            .default = NA)) %>% 
  group_by(genome, environment)
```

```{r microdiversity_core_boxplot,comment="", message=FALSE, warning=FALSE}
ani_environments %>% 
  ggplot(aes(x=environment,y=1-distance,group=environment, color=environment, fill=environment)) +
    geom_jitter(size=1,alpha=0.2, width=0.3) +
    geom_boxplot(outlier.shape = NA) +
    scale_color_manual(values=c("#9E8F71","#429ef5","#f56042"))+
    scale_fill_manual(values=c("#9E8F7170","#429ef570","#f5604270")) +
    ylim(0,0.006) + 
    theme_minimal()
```

```{r microdiversity_core_heatmap,comment="", message=FALSE, warning=FALSE}
ani_pairwise %>% 
  filter(sample.x != sample.y) %>% 
  inner_join(sample_metadata %>% select(sample,environment), by=join_by("sample.x"=="sample")) %>% 
  inner_join(sample_metadata %>% select(sample,environment), by=join_by("sample.y"=="sample")) %>% 
  mutate(environment=case_when(
            environment.x == "low" & environment.y == "high" ~ "different",
            environment.x == "high" & environment.y == "low" ~ "different",
            environment.x == "low" & environment.y == "low" ~ "low",
            environment.x == "high" & environment.y == "high" ~ "high",
            .default = NA)) %>% 
  pairwise_wilcox_test(distance ~ environment, p.adjust.method = "BH")


pairwise_dist <- ani_pairwise %>% 
  group_by(sample.x,sample.y) %>% 
  summarise(distance=mean(distance)) %>% 
  filter(sample.x %in% sample_metadata$sample,
         sample.y %in% sample_metadata$sample) %>% 
  pivot_wider(
      names_from  = sample.y,
      values_from = distance
    ) %>%
  column_to_rownames(var = "sample.x") %>%
  as.matrix() %>%
  { 1 - . } %>% 
  as.dist() 

heatmap_annotation <- sample_metadata %>% 
  filter(sample %in% samples) %>% 
  select(sample,environment,river) %>% 
  column_to_rownames("sample")

pheatmap(
  pairwise_dist,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  clustering_method = "ward.D2",
  display_numbers = FALSE,
  legend = TRUE,
  show_rownames = FALSE,
  show_colnames = FALSE,
  color = colorRampPalette(c("#fbffe0","navy"))(100),
  annotation_col = heatmap_annotation,
  annotation_row = heatmap_annotation,
  main = "Clustered ANI Heatmap"
)
  
```
